[coor.featurizer] split up unit in features and featurizer.
Also raise in case user provided selections/indices are empty. Fixes #549
